feat: Upgrade Vite and React plugins
Updates Vite and its associated React plugin to the latest versions. This includes changes to the vite.config.ts file to use the new plugin and also updates dependencies in package.json and CDN URLs in index.html accordingly. The port configuration in vite.config.ts is also updated to be more flexible.